requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
numpy==1.24.3
outcome==1.2.0
pandas==2.0.1
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
urllib3==2.0.2
wsproto==1.2.0
